Replace Highlight component with react-highlight. Moved contexts into src/contexts
