+++ package.json
@@ -12,12 +12,13 @@
     "bootstrap": "^4.3.1",
     "ci": "^1.0.0",
     "cypress": "^3.4.1",
-    "highlight.js": "^9.15.6",
+    "highlight.js": "^9.15.10",
     "node-sass": "^4.12.0",
     "react": "^16.10.2",
     "react-dom": "^16.10.2",
     "react-ga": "^2.7.0",
     "react-helmet": "^5.2.1",
+    "react-highlight": "^0.12.0",
     "react-router-dom": "^5.1.2",
     "react-scripts": "3.2.0",
     "reactstrap": "^8.0.0"
