+++ src/contexts/index.js
@@ -1,0 +1,4 @@
+import React, { useContext } from "react";
+
+export const Auth0Context = React.createContext();
+export const useAuth0 = () => useContext(Auth0Context);
